Replace test component with real chat UI drawer, lift state to wrapper
