+++ app/dashboard/page.tsx
@@ -1,0 +1,3 @@
+export default function DashboardPage() {
+    return null
+}
